# Optimize SDPA implementation