--- a/candle-core/src/cpu_backend/mod.rs
+++ b/candle-core/src/cpu_backend/mod.rs
@@ -2812,7 +2812,7 @@ impl BackendStorage for CpuStorage {
         MatMul(bmnk).map(self, lhs_l, rhs, rhs_l)
     }
 
-    fn matmul_bias_and_scale(
+    fn matmul_with_beta(
         &self,
         rhs: &Self,
         c: &mut Self,