… .vscode/settings.json
… candle-core/benches/benchmarks/mod.rs
… candle-core/src/dummy_metal_backend.rs
… candle-core/src/error.rs
… candle-core/src/metal_backend/mod.rs
… candle-core/src/quantized/cuda.rs
… candle-core/src/storage.rs
… candle-core/src/tensor.rs
… candle-core/tests/matmul_tests.rs
… candle-metal-kernels/src/lib.rs
… candle-metal-kernels/src/tests.rs
… candle-nn/Cargo.toml
… candle-nn/benches/bench_main.rs
… candle-nn/benches/benchmarks/attention.rs
… candle-nn/benches/benchmarks/mod.rs
… candle-nn/src/attention.rs
… candle-nn/src/cublaslt/api.rs
… candle-nn/src/cublaslt/mod.rs
… candle-nn/src/lib.rs
… candle-transformers/Cargo.toml
… candle-transformers/src/models/mistral.rs
… candle-transformers/src/models/quantized_llama.rs